+++ elfi/examples/lotka_volterra.py
@@ -74,8 +74,10 @@ def lotka_volterra(r1, r2, r3, prey_init=50, predator_init=100, sigma=0., n_obs=
 
     n_full = 20000
     stock = np.empty((batch_size, n_full, 2), dtype=np.int32)
-    stock[:, 0, 0] = prey_init
-    stock[:, 0, 1] = predator_init
+    # As we use approximate continuous priors for prey_init and
+    # predator_init, we'll round them down to closest integers
+    stock[:, 0, 0] = np.floor(prey_init)
+    stock[:, 0, 1] = np.floor(predator_init)
     stoichiometry = np.array([[1, 0], [-1, 1], [0, -1], [0, 0]], dtype=np.int32)
     times = np.empty((batch_size, n_full))
     times[:, 0] = 0
@@ -192,8 +194,8 @@ def get_model(n_obs=50, true_params=None, observation_noise=False, seed_obs=None
         elfi.Prior(ExpUniform, -6., 2., model=m, name='r1'),
         elfi.Prior(ExpUniform, -6., 2., model=m, name='r2'),  # easily kills populations
         elfi.Prior(ExpUniform, -6., 2., model=m, name='r3'),
-        elfi.Prior('poisson', 50, model=m, name='prey0'),
-        elfi.Prior('poisson', 100, model=m, name='predator0')
+        elfi.Prior('normal', 50, np.sqrt(50), model=m, name='prey0'),
+        elfi.Prior('normal', 100, np.sqrt(100), model=m, name='predator0')
     ]
 
     if observation_noise: