+++ CHANGELOG.rst
@@ -1,6 +1,8 @@
 Changelog
 =========
 
+
+- Add boolean `observation_noise` option to `lotka_volterra`
 - Add parameter names as an optional input in model prior and fix the parameter order in priors used in BOLFI and BOLFIRE
 - Add feature names as an optional input and make training data size a required input in BOLFIRE
 - Fix the observed property in simulator nodes