+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+- Fix linting in `arch.py`
 - Add summary statistics to Lotka-Volterra model
 - Add boolean `observation_noise` option to `lotka_volterra`
 - Add parameter names as an optional input in model prior and fix the parameter order in priors used in BOLFI and BOLFIRE