+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 
+- Reformat `summary()` for `Sample(ParameterInferenceResult)`
 - Add boolean `observation_noise` option to `lotka_volterra`
 - Add parameter names as an optional input in model prior and fix the parameter order in priors used in BOLFI and BOLFIRE
 - Add feature names as an optional input and make training data size a required input in BOLFIRE