+++ CHANGELOG.rst
@@ -2,6 +2,7 @@ Changelog
 =========
 
 - Modify Lotka-Volterra model's priors as many methods do not support discrete random variables.
+- Reformat `summary()` for `Sample(ParameterInferenceResult)`
 - Fix linting in `arch.py`
 - Add summary statistics to Lotka-Volterra model
 - Add boolean `observation_noise` option to `lotka_volterra`